Refactor container_fuzzer.go
The fuzzer could share a lot with other integration tests, instead of
downloading containerd 1.5.4.
